+++ tests/Agent/IntegrationTests/IntegrationTests/RequestHeadersCapture/AspNet/AllowAllHeadersEnabledTests.cs
@@ -10,7 +10,7 @@ using NewRelic.Agent.IntegrationTestHelpers.Models;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace NewRelic.Agent.IntegrationTests.RequestHeadersCapture.Asp35
+namespace NewRelic.Agent.IntegrationTests.RequestHeadersCapture.AspNet
 {
     [NetFrameworkTest]
     public class AllowAllHeadersEnabledTests : NewRelicIntegrationTest<RemoteServiceFixtures.BasicMvcApplicationTestFixture>
